Real-client join killers and the serverbound 2168 audit
Two writer bugs a real 1.26.40 client dies on that no roundtrip of BDS
frames could catch, both on branches only our own values take:

- Write(Skin): the zero-animations branch still wrote the count as le32
  where 2168 wants a varint, so every PlayerList add-record desynced the
  client's read cursor by three bytes. ParseSkinColor also overflowed
  int.TryParse for any alpha >= 0x80; parses as uint now.
- McpeNetworkChunkPublisherUpdate never modeled the trailing saved-chunks
  list (plain le32 count + zigzag ChunkPos pairs, in the protocol since
  1.14-era and confirmed by minecraft-data 1001 and gophertunnel). Clients
  through 1.26.33 tolerated the truncation; 1.26.40 does not.

Also carries the serverbound audit against BDS 1.26.40.8: PlayerAuthInput
(flag ordinal list + double-bool optional bodies), ServerBoundDiagnostics
system-categories array, the 2168 item stack variants (li16 descriptor id,
no air short-circuit), the ItemStackRequest action enum shift, and the
client emulator updates that exercised them.

+++ src/MiNET/MiNET.Client/BedrockTraceHandler.cs
@@ -345,7 +345,7 @@ namespace MiNET.Client
 
 			Client.LevelInfo.LevelName = "Default";
 			Client.LevelInfo.Version = 19133;
-			Client.LevelInfo.GameType = message.settings.gameType;
+			Client.LevelInfo.GameType = (int) message.settings.gameType;
 
 			{
 				var packet = McpeRequestChunkRadius.CreateObject();
